requirements.txt
mutagen==1.46.0
yt-dlp==2023.01.06 
python3_discogs_client==2.6
ffmpeg==1.4